requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-dbbackup==4.1.0
django-simple-history==3.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==2.0.5
ecdsa==0.19.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
python-jose==3.3.0
pytz==2024.1
reportlab==4.2.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
